fixed #62
after reloading the search topics from the configuration, the searchtopic panel of the dialog
wasn't listening to the new search topics, but still to the old ones.

now there are new listener added, for the metasearch and the searchtopic.
The dialog is a listener of the metasearch, and listens if new searchtopic are created.
if so, it refreshs its topics panel with the new searchtopics.

the checkboxes of the searchtopicpanel listens to the searchtopics, and is so able to sync
the selection states between the checkboxes and their searchtopics.